# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:5050",
  "servers": [
    {
      "url": "",
      "description": ""
    }
  ],
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/v1/user/": {
      "get": {
        "tags": [
          "User Control"
        ],
        "summary": "Get list of all users in database",
        "description": "",
        "parameters": [
          {
            "name": "phoneNumber",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isSupport",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "list of user",
            "in": "body",
            "type": "array",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "balance": {
                    "type": "number",
                    "example": 0
                  },
                  "role": {
                    "type": "string",
                    "example": "USER"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+911234567891"
                  },
                  "id": {
                    "type": "string",
                    "example": "userid2"
                  },
                  "uid": {
                    "type": "number",
                    "example": 62
                  },
                  "name": {
                    "type": "string",
                    "example": "Dohn Joe"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/v1/user/{id}": {
      "put": {
        "tags": [
          "User Control"
        ],
        "summary": "Initialize a user",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Initialize a user to the database with the users information.",
            "schema": {
              "required": [
                "name"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "description": "name of the user"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Details of the new initialized user",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "username"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+919999999999"
                },
                "role": {
                  "type": "string",
                  "example": "ASTROLOGER"
                },
                "id": {
                  "type": "string",
                  "example": "userid"
                },
                "uid": {
                  "type": "number",
                  "example": 69
                },
                "balance": {
                  "type": "number",
                  "example": 0
                },
                "callStatus": {
                  "type": "string",
                  "example": "online"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User Control"
        ],
        "summary": "Delete a user from database",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "User Control"
        ],
        "summary": "Retrieve a user",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the user",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "username"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+919999999999"
                },
                "role": {
                  "type": "string",
                  "example": "ASTROLOGER"
                },
                "id": {
                  "type": "string",
                  "example": "userid"
                },
                "uid": {
                  "type": "number",
                  "example": 69
                },
                "balance": {
                  "type": "number",
                  "example": 0
                },
                "callStatus": {
                  "type": "string",
                  "example": "online"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User Control"
        ],
        "summary": "Update users call status",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update users call status if ONLIVESTREAM / ONCALL / ONCHAT or ONLINE.",
            "schema": {
              "required": [
                "op",
                "path",
                "value"
              ],
              "properties": {
                "name": {
                  "op": "string",
                  "description": "operation -> REPLACE"
                },
                "path": {
                  "type": "string",
                  "description": "path of the value to change ie status"
                },
                "value": {
                  "type": "string",
                  "description": "the new status ie ONLINE"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Details of the user",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "username"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+919999999999"
                },
                "role": {
                  "type": "string",
                  "example": "ASTROLOGER"
                },
                "id": {
                  "type": "string",
                  "example": "userid"
                },
                "uid": {
                  "type": "number",
                  "example": 69
                },
                "balance": {
                  "type": "number",
                  "example": 0
                },
                "callStatus": {
                  "type": "string",
                  "example": "online"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/v1/user/{id}/settings/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "op": {
                  "example": "any"
                },
                "path": {
                  "example": "any"
                },
                "value": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/user/{id}/session/user/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role": {
                  "example": "any"
                },
                "expiry": {
                  "example": "any"
                },
                "tokentype": {
                  "example": "any"
                },
                "deviceId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/user/{id}/session/user/{id}/{deviceId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/astrologer/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "phoneNumber",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "expertise": {
                  "example": "any"
                },
                "languages": {
                  "example": "any"
                },
                "experience": {
                  "example": "any"
                },
                "aboutMe": {
                  "example": "any"
                },
                "phoneNumber": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/astrologer/{id}": {
      "get": {
        "tags": [
          "Host Profile Control"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "op": {
                  "example": "any"
                },
                "path": {
                  "example": "any"
                },
                "value": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/astrologer/{id}/settings/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "op": {
                  "example": "any"
                },
                "path": {
                  "example": "any"
                },
                "value": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/astrologer/{id}/profile/review/": {
      "get": {
        "tags": [
          "User Review Control"
        ],
        "summary": "Get all reviews from startTs to endTs ",
        "description": "Require hostId, startTs, endTs and key\\n ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "startTs",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "starting ts to get reviews from"
          },
          {
            "name": "hostId",
            "required": true,
            "type": "string",
            "description": "id of the host",
            "in": "query"
          },
          {
            "name": "endTs",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "end timestamp till where to get reviews"
          },
          {
            "name": "key",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "the key from which to get the left reviews from, used for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Review data ",
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "isan"
                },
                "comment": {
                  "type": "string",
                  "example": "hi there"
                },
                "reply": {
                  "type": "string",
                  "example": ""
                },
                "hostId": {
                  "type": "string",
                  "example": "01J6VDPDEGZN7DED7TAWFQ3NWK"
                },
                "rating": {
                  "type": "number",
                  "example": 3
                },
                "tsUserId": {
                  "type": "string",
                  "example": "1731059521080#01HV9QW1W342YY2T6C4N7NVSRM"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User Review Control"
        ],
        "summary": "Add a review for a host",
        "description": "Require userId, and amount\\n ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "hostId",
            "required": true,
            "type": "string",
            "description": "id of the host",
            "in": "query"
          },
          {
            "name": "body",
            "in": "body",
            "description": "userId and amount.",
            "schema": {
              "required": [
                "userId",
                "userName",
                "rating",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "description": "comment"
                },
                "userId": {
                  "type": "string",
                  "description": "userId of the user"
                },
                "userName": {
                  "type": "string",
                  "description": "userName of the user"
                },
                "rating": {
                  "type": "number",
                  "description": "rating to give to the host"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review data ",
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "isan"
                },
                "comment": {
                  "type": "string",
                  "example": "hi there"
                },
                "reply": {
                  "type": "string",
                  "example": ""
                },
                "hostId": {
                  "type": "string",
                  "example": "01J6VDPDEGZN7DED7TAWFQ3NWK"
                },
                "rating": {
                  "type": "number",
                  "example": 3
                },
                "tsUserId": {
                  "type": "string",
                  "example": "1731059521080#01HV9QW1W342YY2T6C4N7NVSRM"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/v1/astrologer-view/{id}": {
      "get": {
        "tags": [
          "Host Profile Control"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-ASTRO-CONTEXT",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/astrologer-view/search/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isOnline",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortCategory",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "type": "string"
          },
          {
            "name": "expertise",
            "in": "query",
            "type": "string"
          },
          {
            "name": "language",
            "in": "query",
            "type": "string"
          },
          {
            "name": "gender",
            "in": "query",
            "type": "string"
          },
          {
            "name": "topAstrologer",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/astrologer-view/{id}/settings/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "op": {
                  "example": "any"
                },
                "path": {
                  "example": "any"
                },
                "value": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/astrologer-view/{id}/profile/review/": {
      "get": {
        "tags": [
          "User Review Control"
        ],
        "summary": "Get all reviews from startTs to endTs ",
        "description": "Require hostId, startTs, endTs and key\\n ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "startTs",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "starting ts to get reviews from"
          },
          {
            "name": "hostId",
            "required": true,
            "type": "string",
            "description": "id of the host",
            "in": "query"
          },
          {
            "name": "endTs",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "end timestamp till where to get reviews"
          },
          {
            "name": "key",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "the key from which to get the left reviews from, used for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Review data ",
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "isan"
                },
                "comment": {
                  "type": "string",
                  "example": "hi there"
                },
                "reply": {
                  "type": "string",
                  "example": ""
                },
                "hostId": {
                  "type": "string",
                  "example": "01J6VDPDEGZN7DED7TAWFQ3NWK"
                },
                "rating": {
                  "type": "number",
                  "example": 3
                },
                "tsUserId": {
                  "type": "string",
                  "example": "1731059521080#01HV9QW1W342YY2T6C4N7NVSRM"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User Review Control"
        ],
        "summary": "Add a review for a host",
        "description": "Require userId, and amount\\n ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "hostId",
            "required": true,
            "type": "string",
            "description": "id of the host",
            "in": "query"
          },
          {
            "name": "body",
            "in": "body",
            "description": "userId and amount.",
            "schema": {
              "required": [
                "userId",
                "userName",
                "rating",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "description": "comment"
                },
                "userId": {
                  "type": "string",
                  "description": "userId of the user"
                },
                "userName": {
                  "type": "string",
                  "description": "userName of the user"
                },
                "rating": {
                  "type": "number",
                  "description": "rating to give to the host"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review data ",
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "isan"
                },
                "comment": {
                  "type": "string",
                  "example": "hi there"
                },
                "reply": {
                  "type": "string",
                  "example": ""
                },
                "hostId": {
                  "type": "string",
                  "example": "01J6VDPDEGZN7DED7TAWFQ3NWK"
                },
                "rating": {
                  "type": "number",
                  "example": 3
                },
                "tsUserId": {
                  "type": "string",
                  "example": "1731059521080#01HV9QW1W342YY2T6C4N7NVSRM"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/v1/chat/{id}/user/{userId}": {
      "patch": {
        "tags": [
          "Chat Control"
        ],
        "summary": "Modify users lastRead and lastReceived message timestamp for chat",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identifier of the chat session"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identifier of the user for modifying lastReceived or lastRead"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update waitlist of a livestream using new users information",
            "schema": {
              "required": [
                "op",
                "path",
                "value"
              ],
              "properties": {
                "op": {
                  "type": "string",
                  "description": "operation for patch request: {REPLACE}"
                },
                "path": {
                  "type": "string",
                  "description": "value to change {READ, RECEIVED} "
                },
                "value": {
                  "type": "number",
                  "description": "value to be assigned"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat session",
            "schema": {
              "type": "object",
              "properties": {
                "userList": {
                  "type": "string",
                  "example": "01HPGHBN8NWGG4VFNP3PM4ND6N#01HPKDB7ENB0FDEGC62DKK00HR"
                },
                "users": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "lastRead": {
                        "type": "number",
                        "example": 1712659817975
                      },
                      "lastReceived": {
                        "type": "number",
                        "example": 1712659817975
                      },
                      "id": {
                        "type": "string",
                        "example": "01HPKDB7ENB0FDEGC62DKK00HR"
                      },
                      "role": {
                        "type": "string",
                        "example": "USER"
                      }
                    }
                  }
                },
                "id": {
                  "type": "string",
                  "example": "01HV17NSFQH0J2ER62SC6H6F2M"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/v1/support-case/user/{userId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/session/user/{id}/{deviceId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}